New attributes added to the card model
Added arrays to hold participants (users assigned, or watching, cards) and labels. Replaced created_at with date, since its more user friendly (because it is shown in the yaml frontmatter), and added a new hash to store additional properties which the user may need to store (such as, for example, priority).